blog_posts/requirements.txt
django_tinymce==3.3.0
Django==3.2.5
